+++ src/main.cpp
@@ -1,0 +1,11 @@
+#include <iostream>
+#include <QDebug>
+
+int main(int argc, char* argv[])
+{
+    std::cout << "Hello!" << std::endl;
+
+    qInfo() << "Hello from Qt!";
+
+    return 0;
+}